Refactor: Move MainActivity to main_activity package
