--- app/src/main/java/com/giraffe/cineverseapp/MainActivity.kt
@@ -1,39 +1,0 @@
-package com.giraffe.cineverseapp
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.SystemBarStyle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
-import com.giraffe.authentication.AuthenticationApi
-import com.giraffe.designsystem.theme.CineVerseTheme
-import dagger.hilt.android.AndroidEntryPoint
-import jakarta.inject.Inject
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    @Inject
-    lateinit var authenticationApi: AuthenticationApi
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge(
-            statusBarStyle = SystemBarStyle.auto(
-                Color.Transparent.toArgb(),
-                Color.Transparent.toArgb()
-            ),
-            navigationBarStyle = SystemBarStyle.auto(
-                Color.Transparent.toArgb(),
-                Color.Transparent.toArgb()
-            )
-        )
-        setContent {
-            CineVerseTheme {
-                authenticationApi.LoginContainer { }
-            }
-        }
-    }
-}
